[#2] refactor: MdContentService 추가
